internal: drop SharedData
